+++ package.json
@@ -22,9 +22,9 @@
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.2",
     "@react-icons/all-files": "^4.1.0",
-    "@rjsf/core": "^5.19.4",
-    "@rjsf/utils": "^5.19.4",
-    "@rjsf/validator-ajv8": "^5.19.4",
+    "@rjsf/core": "^5.20.0",
+    "@rjsf/utils": "^5.20.0",
+    "@rjsf/validator-ajv8": "^5.20.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
@@ -39,15 +39,15 @@
     "react-syntax-highlighter": "^15.5.0",
     "react-wrap-balancer": "^1.1.1",
     "sort-json": "^2.0.1",
-    "tailwind-merge": "^2.4.0",
+    "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^0.9.1",
     "yaml": "^2.5.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@types/node": "^20.14.14",
-    "@types/react": "^18.3.3",
+    "@types/node": "^20.16.1",
+    "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
     "@types/react-syntax-highlighter": "^15.5.13",
     "@types/sort-json": "^2.0.3",
@@ -55,11 +55,11 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "14.1.0",
     "eslint-config-prettier": "^9.1.0",
-    "postcss": "^8.4.40",
+    "postcss": "^8.4.41",
     "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.5",
-    "shiki": "^1.12.1",
-    "tailwindcss": "^3.4.7",
+    "prettier-plugin-tailwindcss": "^0.6.6",
+    "shiki": "^1.14.1",
+    "tailwindcss": "^3.4.10",
     "typescript": "^5.5.4"
   }
 }
