+++ src/app/globals.css
@@ -92,3 +92,18 @@
   background-image: radial-gradient(#1e293b30 0.8px, #fff 0.8px);
   background-size: 24px 24px;
 }
+
+/*
+
+.rjsf {
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  display: grid
+}
+
+.form_group {
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  display: grid
+}
+
+
+`*/
