+++ src/components/form/cluster.tsx
@@ -5,7 +5,6 @@ import React from "react";
 import validator from "@rjsf/validator-ajv8";
 import { stringify } from "yaml";
 import SyntaxHighlighter from "react-syntax-highlighter";
-import { uiSchema } from "@/components/form/uischema";
 
 import SubmitButton from "./custom/SubmitButton";
 
@@ -41,20 +40,93 @@ import {
   SelectLabel,
 } from "@/components/ui/select";
 
-import { getDefaultRegistry } from "@rjsf/core";
-import { ObjectFieldTemplateProps } from "@rjsf/utils";
-const registry = getDefaultRegistry();
-
-const ObjectFieldTemplate = registry.templates.ObjectFieldTemplate;
+const uiSchema: UiSchema = {
+  kind: { "ui:widget": "hidden" },
+  apiVersion: { "ui:widget": "hidden" },
+  metadata: {
+    name: {
+      "ui:title": "Cluster name",
+      "ui:description": "Name of your Cluster",
+    },
+    namespace: {
+      "ui:description": "Namespace in which to deploy the Cluster",
+    },
+    labels: { "ui:widget": "hidden" },
+  },
+  spec: {
+    clusterNetwork: {
+      serviceDomain: { "ui:widget": "hidden" },
+      services: {
+        cidrBlocks: {
+          items: {
+            "ui:title": "Service CIDRs",
+            "ui:description": "CIDR addresses for all services in the cluster",
+          },
+        },
+      },
+      pods: {
+        cidrBlocks: {
+          items: {
+            "ui:title": "Pod CIDRs",
+            "ui:description": "CIDR addresses for all pods in the cluster",
+          },
+        },
+      },
+    },
+    topology: {
+      class: { "ui:widget": "hidden" },
+      version: {
+        "ui:title": "Kubernetes version",
+        "ui:description": "Select which Kubernetes minor version you want",
+        "ui:field": "k8s_version",
+      },
+      controlPlane: {
+        replicas: {
+          "ui:title": "Control plane replicas",
+          "ui:description": "How many control plane replicas you want",
+          "ui:widget": "updown",
+        },
+      },
+      workers: {
+        machineDeployments: {
+          items: {
+            name: { "ui:widget": "hidden" },
+            failureDomain: { "ui:widget": "hidden" },
+            class: { "ui:widget": "hidden" },
+            replicas: {
+              "ui:title": "Worker replicas",
+              "ui:description": "How many worker replicas you want",
+              "ui:widget": "updown",
+            },
+          },
+        },
+      },
+    },
+  },
+};
 
 export const ClusterForm = (schema: any) => {
-  const schemas = schema?.schema;
-  const list = Object.keys(schemas).reverse();
+  const cluster_schema = schema?.schema;
+  const list = Object.keys(cluster_schema).reverse();
+  //console.log(schema)
 
-  console.log(
-    schemas["openstack-scs-1-30-v1"].properties.spec.properties.topology
-      .properties.variables,
-  );
+  function removeVariables(schema: any) {
+    const newSchema = JSON.parse(JSON.stringify(schema));
+
+    if (
+      newSchema.properties &&
+      newSchema.properties.spec &&
+      newSchema.properties.spec.properties &&
+      newSchema.properties.spec.properties.topology &&
+      newSchema.properties.spec.properties.topology.properties
+    ) {
+      delete newSchema.properties.spec.properties.topology.properties.variables;
+    }
+
+    return newSchema;
+  }
+
+  const modifiedSchema = removeVariables(cluster_schema);
 
   // Define the custom field components to register; here our "geo"
   // custom field component
@@ -62,7 +134,9 @@ export const ClusterForm = (schema: any) => {
 
   const [clusterstack, setClusterStack] = useState("openstack-scs-1-30-v1");
   const [formData, setFormData] = useState(null);
-  const [activeSchema, setActiveSchema] = useState(schemas[clusterstack]);
+  const [activeSchema, setActiveSchema] = useState(
+    cluster_schema[clusterstack],
+  );
 
   const handleSwitch = (value: string) => {
     setClusterStack(value);
@@ -72,141 +146,35 @@ export const ClusterForm = (schema: any) => {
 
   const yaml_out = stringify(formData).trimEnd(); // json to yaml conversion
 
-  // CUSTOM GROUPS (Cluster, Machines, Variables)
-
-  const groups = [
-    {
-      title: "cluster",
-      fields: [
-        "metadata.properties.name",
-        "metadata.properties.namespace",
-        "version",
-      ],
-    },
-    { title: "machines", fields: "spec.variables" },
-    //{ title: "variables", fields: "spec.variables" },
-  ];
-
-  const getPropsForGroup = (
-    group: any,
-    props: ObjectFieldTemplateProps,
-  ): ObjectFieldTemplateProps => {
-    2;
-    console.log(props.properties);
-    console.log(group.fields);
-    console.log(props.properties.filter((p) => console.log(p)));
-
-    // More filtering might be required for propper functionality, this is just a POC
-    return {
-      ...props,
-      // properties: props.properties.filter((p) => group.fields.includes(p.name)),
-    };
-  };
-
-  const ObjectFieldTemplateWrapper = (props: ObjectFieldTemplateProps) => {
-    return (
-      <>
-        {groups.map((group) => {
-          const childProps = getPropsForGroup(group, props);
-          console.log(group);
-          return (
-            <>
-              <Card>
-                <CardHeader>
-                  <CardTitle className="">
-                    {group.title.charAt(0).toUpperCase() + group.title.slice(1)}
-                  </CardTitle>
-                </CardHeader>
-                <CardContent className="space-y-4">
-                  <div>
-                    <ObjectFieldTemplate key={group.title} {...childProps} />
-                  </div>
-                </CardContent>
-              </Card>
-            </>
-          );
-        })}
-      </>
-    );
-  };
-
   return (
     <>
-      <div className="flex space-x-8 mt-8">
-        <div className="grid grid-row-cols grid-cols-2 gap-12 mb-4">
-          <div className="">
-            <div className="mb-4">
-              <Card>
-                <CardHeader>
-                  <CardTitle className="">Cluster Stack</CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <Select
-                    onValueChange={(value) => handleSwitch(value)}
-                    defaultValue={clusterstack}
-                  >
-                    <SelectTrigger>
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {list.map((item) => (
-                        <SelectItem
-                          key={item}
-                          value={item.toString()}
-                          defaultValue="openstack-scs-1-30-v1"
-                        >
-                          {item}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                  <p className="text-sm text-muted-foreground mt-2">
-                    Select which Cluster Stack to use
-                  </p>
-                </CardContent>
-              </Card>
-            </div>
-
-            <div>
-              <RJSForm
-                schema={activeSchema}
-                uiSchema={uiSchema}
-                //formData={formData}
-                validator={validator}
-                fields={fields}
-                widgets={widgets}
-                onChange={(e) => setFormData(e.formData)}
-                templates={{
-                  ButtonTemplates: { SubmitButton },
-                  //ObjectFieldTemplate: ObjectFieldTemplateWrapper,
-                }}
-              >
-                {/*<DownloadButton formStatus={validator} formData={formData}/> */}
-              </RJSForm>
-
-              {/* <MultiInput value={values} onchange={setValues} /> */}
-            </div>
+      <div className="space-x-8 mt-8">
+        <div className=" mb-4">
+          <div>
+            <Card>
+              <CardHeader>
+                <CardTitle className="">Cluster</CardTitle>
+              </CardHeader>
+              <CardContent className="space-y-4">
+                <RJSForm
+                  schema={modifiedSchema}
+                  uiSchema={uiSchema}
+                  //formData={formData}
+                  validator={validator}
+                  fields={fields}
+                  widgets={widgets}
+                  onChange={(e) => setFormData(e.formData)}
+                  templates={{
+                    ButtonTemplates: { SubmitButton },
+                    //ObjectFieldTemplate: ObjectFieldTemplateWrapper,
+                  }}
+                >
+                  {/*<DownloadButton formStatus={validator} formData={formData}/> */}
+                </RJSForm>
+              </CardContent>
+            </Card>
+            {/* <MultiInput value={values} onchange={setValues} /> */}
           </div>
-          <SyntaxHighlighter
-            language="yaml"
-            className="text-sm w-1/2"
-            style={{
-              "hljs-attr": {
-                color: "hsl(var(--foreground))",
-              },
-              "react-syntax-highlighter-line-number": {
-                color: "rgb(183 183 183)",
-                margin: "0",
-              },
-            }}
-            showLineNumbers={true}
-            customStyle={{
-              backgroundColor: "var(--primary-color)",
-              color: "#0F5FE1",
-            }}
-          >
-            {yaml_out}
-          </SyntaxHighlighter>
         </div>
       </div>
     </>
